misc updates
Updated angular versions
Updated readme
Some folder restructuring/organizing
