+++ TextMaker.api/ClientModels/Project.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace TextMaker
+{
+	[Serializable]
+	public class Project
+	{
+		public string Name { get; set; }
+
+		public TextItemGroup[] Groups { get; set; }
+
+		public bool IsDateSelected { get; set; }
+
+		public bool iIsAutomaticallyCopied { get; set; }
+	}
+}
