+++ TextMaker.api/appsettings.json
@@ -1,0 +1,10 @@
+{
+  "Logging": {
+      "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+      }
+    },
+"AllowedHosts": "*"
+}
