+++ TextMaker.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29609.76
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TextMaker", "TextMaker\TextMaker.csproj", "{C347EA96-9C2C-40D8-B34A-B295E3D28550}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TextMaker", "TextMaker.api\TextMaker.csproj", "{C347EA96-9C2C-40D8-B34A-B295E3D28550}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
